updated read me file & loan offer calculation function

+++ src/main/java/com/loanlender/app/dto/LoanDetails.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class LoanDetails {
     private int loanId;
     private double tenureRate;
-    private int tenure;
+    private double tenure;
     private double dailyPayment;
     private double totalPayment;
     private double loanAmount;
